Fixed a bug with 5.0.2 regarding issuing and parsing selenium commands.
do-command send/receive was failing due to a timing issue: bytes-ready? was being called before any data had hit the response port. A workaround was to use read-line, which waits for data to appear in the port, manually parse the headers and then use the existing parsing procedure with the remaining data on the port.